+++ apps/dl-smudge.el
@@ -1,10 +1,11 @@
-;;; dl-spotify.el --- Spotify :: Smudge -*- lexical-binding: t; -*-
+;;; dl-smudge.el --- Spotify :: Smudge -*- lexical-binding: t; -*-
 
 (load (expand-file-name
         "dl-spotify.secret.el" user-emacs-directory) t)
 
 (setq smudge-transport 'connect)
 (use-package smudge
+  :defer t
   :custom
   (smudge-oauth2-client-secret dl/spotify-client-id)
   (smudge-oauth2-client-id dl/spotify-client-id)
@@ -15,5 +16,6 @@
   (global-smudge-remote-mode)
   :bind (("C-c ." . smudge-command-map)))
 
-(provide 'dl-spotify)
-;;; dl-spotify.el ends here
+
+(provide 'dl-smudge)
+;;; dl-smudge.el ends here