+++ apps/dl-treemacs.el
@@ -1,5 +1,0 @@
-;;; dl-treemacs.el --- Treemacs file sidebar -*- lexical-binding: t; -*-
-
-(use-package treemacs)
-(provide 'dl-treemacs)
-;;; dl-treemacs.el ends here