+++ completion/dl-vertico.el
@@ -1,4 +1,4 @@
-;;; dl-orderless.el --- Vertico config -*- lexical-binding: t; -*-
+;;; dl-vertico.el --- Vertico config -*- lexical-binding: t; -*-
 
 ;; lots here https://github.com/minad/vertico/tree/main
 
@@ -16,26 +16,8 @@
 ;;   (use-package vertico-posframe)
 ;;   (vertico-posframe-mode nil))
 
-;; Persist history over Emacs restarts. Vertico sorts by history position.
-(use-package savehist
-  :init
-  (savehist-mode))
-
-;; Emacs minibuffer configurations.
-(use-package emacs
-  :custom
-  ;; Enable context menu. `vertico-multiform-mode' adds a menu in the minibuffer
-  ;; to switch display modes.
-  (context-menu-mode t)
-  ;; Support opening new minibuffers from inside existing minibuffers.
-  (enable-recursive-minibuffers t)
-  ;; Hide commands in M-x which do not work in the current mode.  Vertico
-  ;; commands are hidden in normal buffers. This setting is useful beyond
-  ;; Vertico.
-  (read-extended-command-predicate #'command-completion-default-include-p)
-  ;; Do not allow the cursor in the minibuffer prompt
-  (minibuffer-prompt-properties
-    '(read-only t cursor-intangible t face minibuffer-prompt)))
+;; Minibuffer-wide settings (recursive minibuffers, prompt properties,
+;; M-x predicate, context-menu) live in `dl-completion.el' / `dl-interface.el'.
 
 (keymap-set vertico-map "M-?"   #'minibuffer-completion-help)
 (keymap-set vertico-map "M-RET" #'minibuffer-force-complete-and-exit)