+++ lang/dl-lang-common.el
@@ -6,9 +6,6 @@
 ;;;
 ;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
 
-(use-package markdown-mode
-  :hook ((markdown-mode . visual-line-mode)))
-
 (use-package yaml-mode)
 
 (use-package json-mode)